+++ R/tcn_counts.R
@@ -10,9 +10,11 @@
 #'   default to 30 days before end_time. Default is NULL.
 #' @param end_time Character string. Latest tweet timestamp to return (UTC in ISO 8601 format). If NULL API will default
 #'   to now - 30 seconds. Default is NULL.
-#' @param granularity Character string. Granularity or period for tweet counts. Can be "day", "minute" or "hour". Default is "day".
+#' @param granularity Character string. Granularity or period for tweet counts. Can be "day", "minute" or "hour".
+#'   Default is "day".
 #' @param retry_on_limit Logical. When the API v2 rate-limit has been reached wait for reset time. Default
 #'   is TRUE.
+#' @param verbose Logical. Output additional information. Default is \code{TRUE}.
 #'
 #' @note A rate-limit of 300 requests per 15 minute window applies. If a conversation count request contains more than
 #'   31 days-worth of results it will use more than one request as API results will be paginated.
@@ -44,7 +46,10 @@ tcn_counts <-
            start_time = NULL,
            end_time = NULL,
            granularity = "day",
-           retry_on_limit = TRUE) {
+           retry_on_limit = TRUE,
+           verbose = TRUE) {
+
+    msg <- f_verbose(verbose)
 
     # check params
     if (is.null(token$bearer) || !is.character(token$bearer)) {
@@ -79,8 +84,12 @@ tcn_counts <-
 
     results <- lst_counts_results()
 
-    pb <- prog_bar(length(ids), paste0(" of ", length(ids), " | counts"))
-    pb$tick(0)
+    msg(paste0("voson.tcn conversation counts: n=", length(ids), ", min-api-requests=", length(ids), ", endpoint=", endpoint))
+
+    if (verbose) {
+      pb <- prog_bar(length(ids), "")
+      pb$tick(0)
+    }
 
     for (convo_id in ids) {
       # endpoint description
@@ -92,10 +101,10 @@ tcn_counts <-
 
       # check rate-limit
       if (resp$status == 429) {
-        message(paste0("\ntwitter api rate-limit reached at ", Sys.time()))
+        msg(paste0("\ntwitter api rate-limit reached at ", Sys.time()))
         reset <- resp$headers$`x-rate-limit-reset`
         if (retry_on_limit & !is.null(reset)) {
-          rl_status <- resp_rate_limit(resp$headers, endpoint_desc, TRUE)
+          rl_status <- resp_rate_limit(resp$headers, endpoint_desc, sleep = TRUE, verbose = verbose)
 
           # repeat request after reset
           if (rl_status) {
@@ -104,7 +113,7 @@ tcn_counts <-
             next_token <- NULL
           }
         } else {
-          rl_status <- resp_rate_limit(resp$headers, endpoint_desc, FALSE)
+          rl_status <- resp_rate_limit(resp$headers, endpoint_desc, sleep = FALSE, verbose = verbose)
           next_token <- NULL
           break
         }
@@ -123,9 +132,10 @@ tcn_counts <-
         # will not occur for recent endpoint as it only retrieves last 7 days
         next_token <- resp_data$meta[["next_token"]]
       } else {
-        message(
+        err_msg(
           paste0(
-            "\ntwitter api response status (", endpoint_desc, "): ", resp$status, "\n",
+            "\nresponse error:\n",
+            "api status (", endpoint_desc, "): ", resp$status, "\n",
             "conversation_id: ", convo_id, ", next_token: -"
           ))
         next_token <- NULL
@@ -145,10 +155,10 @@ tcn_counts <-
 
         # check rate-limit
         if (resp$status == 429) {
-          message(paste0("\ntwitter api rate-limit reached at ", Sys.time()))
+          msg(paste0("\ntwitter api rate-limit reached at ", Sys.time()))
           reset <- resp$headers$`x-rate-limit-reset`
           if (retry_on_limit & !is.null(reset)) {
-            rl_status <- resp_rate_limit(resp$headers, endpoint_desc, sleep = TRUE)
+            rl_status <- resp_rate_limit(resp$headers, endpoint_desc, sleep = TRUE, verbose = verbose)
 
             # repeat request after reset
             if (rl_status) {
@@ -157,7 +167,7 @@ tcn_counts <-
               next_token <- NULL
             }
           } else {
-            rl_status <- resp_rate_limit(resp$headers, endpoint_desc, sleep = FALSE)
+            rl_status <- resp_rate_limit(resp$headers, endpoint_desc, sleep = FALSE, verbose = verbose)
             next_token <- NULL
           }
         }
@@ -173,9 +183,10 @@ tcn_counts <-
 
           next_token <- resp_data$meta[["next_token"]]
         } else {
-          message(
+          err_msg(
             paste0(
-              "\ntwitter api response status (", endpoint_desc, "): ", resp$status, "\n",
+              "\nresponse error:\n",
+              "api status (", endpoint_desc, "): ", resp$status, "\n",
               "conversation_id: ", convo_id, ", next_token: ", next_token
             ))
           next_token <- NULL
@@ -184,7 +195,7 @@ tcn_counts <-
         req_ts <- as.integer(as.POSIXct(Sys.time()))
       }
 
-      pb$tick(1)
+      if (verbose) pb$tick(1)
     }
 
     results
@@ -198,7 +209,7 @@ resp_content_counts <- function(resp) {
   content <- tryCatch({
     jsonlite::fromJSON(httr::content(resp, as = "text", encoding = "UTF-8"), flatten = TRUE)
   }, error = function(e) {
-    message("JSON content error: ", e)
+    err_msg(paste0("JSON content error: ", e))
     return(NULL)
   })