+++ R/utils.R
@@ -31,9 +31,11 @@ ids_from_urls <- function(x) {
 }
 
 # api rate-limit
-resp_rate_limit <- function(headers, endpoint = NULL, sleep = FALSE) {
+resp_rate_limit <- function(headers, endpoint = NULL, sleep = FALSE, verbose = TRUE) {
+  msg <- f_verbose(verbose)
+
   reset <- as.numeric(headers$`x-rate-limit-reset`)
-  message(
+  msg(
     paste0(
       ifelse(is.null(endpoint), "", paste0(endpoint, " ")),
       "remaining: ",
@@ -56,7 +58,7 @@ resp_rate_limit <- function(headers, endpoint = NULL, sleep = FALSE) {
       return(FALSE)
     }
 
-    message(paste0("waiting ", round((wait_time/60), digits = 1), " mins for reset"))
+    msg(paste0("waiting ", round((wait_time/60), digits = 1), " mins for reset"))
     Sys.sleep(wait_time + 2) # add 2 sec buffer
   }
 
@@ -123,38 +125,99 @@ clean_results <- function(results) {
 
 # build conversation id search queries based on ids and endpoint
 build_chunks <- function(ids, endpoint = "recent") {
-  if (length(ids) == 1) return(paste0("conversation_id:", ids))
-
   q_size <- 512
   if (endpoint == "all") q_size <- 1024
+
+  # remove trailing "or" and encode "or" separators then add chunk
+  add_chunk <- function(chunks, x) {
+    chunk_qs <- sub("\\sOR\\s$", "", x)
+    chunk_qs <- gsub("\\sOR\\s", "%20OR%20", chunk_qs)
+    c(chunks, chunk_qs)
+  }
 
   chunks <- c()
   qs <- ""
   n_chr_tally <- 0
   for (x in ids) {
     q <- paste0("conversation_id:", x)
+
     if ((n_chr_tally + nchar(q)) > q_size) {
-      qs <- sub("\\sOR\\s$", "", qs)
-      qs <- gsub("\\sOR\\s", "%20OR%20", qs)
-      chunks <- c(chunks, qs)
+      chunks <- add_chunk(chunks, qs)
+
       qs <- ""
       n_chr_tally <- 0
-    } else {
-      qs <- paste0(qs, q, " OR ")
-      n_chr_tally <- nchar(qs)
     }
+
+    qs <- paste0(qs, q, " OR ")
+    n_chr_tally <- nchar(qs)
   }
+
+  if (nchar(qs) > 0) chunks <- add_chunk(chunks, qs)
 
   chunks
 }
 
+# progress bar format
 prog_bar <- function(n_iter, desc) {
   progress::progress_bar$new(
-    format = paste0("[:bar] :percent ", desc),
+    format = paste0("downloading [:bar] :percent", desc),
     total = n_iter,
     complete = "=",
     incomplete = "-",
     current = ">",
-    width = 60)
+    width = 60,
+    clear = FALSE
+  )
 }
 
+# assign message function for verbose output
+f_verbose <- function(x) {
+  opt_msg <- getOption("voson.msg")
+  opt_f <- vtcn_msg
+  if (!is.null(opt_msg) && !is.na(opt_msg)) {
+    if (is.logical(opt_msg)) {
+      if (opt_msg == FALSE) {
+        opt_f <- vtcn_cat
+      }
+    }
+  }
+
+  f <- vtcn_silent
+  if (!is.null(x)) {
+    if (is.logical(x)) {
+      if (x) f <- opt_f
+    }
+  }
+
+  f
+}
+
+# errors
+err_msg <- function(x) {
+  opt_msg <- getOption("voson.msg")
+  if (!is.null(opt_msg) && !is.na(opt_msg)) {
+    if (is.logical(opt_msg)) {
+      if (opt_msg == FALSE) {
+        vtcn_cat(x)
+        return()
+      }
+    }
+  }
+  invisible(vtcn_msg(x))
+}
+
+# base output function
+msg <- function(x) vtcn_msg(x)
+
+# message output
+vtcn_msg <- function(x) message(x) # appendLF = FALSE
+
+# cat output
+vtcn_cat <- function(x) {
+  cat(x)
+  flush.console()
+  Sys.sleep(0.05)
+}
+
+# silent output
+vtcn_silent <- function(x) return()